Add shared class for overriding registered install registry key in tests
