+++ src/test/HostActivationTests/MultilevelSDKLookup.cs
@@ -566,30 +566,10 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
 
             var dotnet = fixture.BuiltDotnet;
 
-            // To correctly test the product we need a registry key which is
-            // - writable without admin access (so that the tests don't require admin to run)
-            // - redirected in WOW64 - so that there are both 32bit and 64bit versions of the key
-            //   this is because the product stores the info in the 32bit version only and even 64bit
-            //   product must look into the 32bit version.
-            //   Without the redirection we would not be able to test that the product always looks
-            //   into 32bit only.
-            // Per this page https://docs.microsoft.com/en-us/windows/desktop/WinProg64/shared-registry-keys
-            // a user writable redirected key is for example HKCU\Software\Classes\Interface
-            // so we're going to use that one - it's not super clean as they key stored COM interfaces
-            // but we should not corrupt anything by adding a special subkey even if it's left behind.
-            //
-            // Note: If you want to inspect the values written by the test and/or modify them manually
-            //   you have to navigate to HKCU\Software\Classes\Wow6432Node\Interface on a 64bit OS.
-
-            RegistryKey hkcu = RegistryKey.OpenBaseKey(RegistryHive.CurrentUser, RegistryView.Registry32);
-            RegistryKey interfaceKey = hkcu.CreateSubKey(@"Software\Classes\Interface");
-            string testKeyName = "_DOTNET_Test" + System.Diagnostics.Process.GetCurrentProcess().Id.ToString();
-            RegistryKey testKey = interfaceKey.CreateSubKey(testKeyName);
-            try
+            using (var regKeyOverride = new RegisteredInstallKeyOverride())
             {
                 string architecture = fixture.CurrentRid.Split('-')[1];
-                RegistryKey dotnetLocationKey = testKey.CreateSubKey($@"Setup\InstalledVersions\{architecture}");
-                dotnetLocationKey.SetValue("InstallLocation", _regDir);
+                regKeyOverride.SetInstallLocation(_regDir, architecture);
 
                 // Add SDK versions
                 AddAvailableSdkVersions(_regSdkBaseDir, "9999.0.4");
@@ -605,16 +585,12 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                     .WithUserProfile(_userDir)
                     .Environment(s_DefaultEnvironment)
                     .EnvironmentVariable("DOTNET_MULTILEVEL_LOOKUP", "1")
-                    .EnvironmentVariable("_DOTNET_TEST_SDK_REGISTRY_PATH", testKey.Name)
+                    .EnvironmentVariable("_DOTNET_TEST_SDK_REGISTRY_PATH", regKeyOverride.KeyPath)
                     .CaptureStdOut()
                     .CaptureStdErr()
                     .Execute()
                     .Should().Pass()
                     .And.HaveStdErrContaining(Path.Combine(_regSelectedMessage, "9999.0.4", _dotnetSdkDllMessageTerminator));
-            }
-            finally
-            {
-                interfaceKey.DeleteSubKeyTree(testKeyName);
             }
         }
